Adjusted the win32 'platform build' (which creates an old-school Contiki OS) to the changed dependency handling.
The special module-compile rule didn't work anymore now that the make doesn't create the .d beforehand. However the ordinary compile rule in Makefile.include turns out to create object files linkable as DLLs - although with warnings. The downside of these warnings seems to be outweight by the benefit of a simpler Makefile.