Fixed recent build error in example-shell.
Both apps/webbrowser and apps/webserver contain a http-strings.c. It seems unclear to me if the original intention was to have them identical (but then they should have been factored out in the first place) or if they were only very similiar by chance.

Anyway, currently webserver/http-strings.c is a clean superset of webbrowser/http-strings.c so if a project has both HTTP server and client parts it is desirable to use the webserver variant. In the case of apps/shell this can be archived by adding webserver *before* webbrowser to the APPS variable.

This seems like a hack to me - but the whole shell build qualifies as hack, doesn't it ;-)